Fix build, build with ng-packagr, component cleanup, update to 1.0.6
