+++ .gitignore
@@ -1,8 +1,1 @@
-/nbproject/private/
 /node_modules
-/build
-/nbproject
-
-/node_modules
-/build
-/nbproject
