--- config-library.js
@@ -1,3 +1,0 @@
-export const nameLibrary = "ngu-utility";
-export const PATH_SRC = "lib/";
-export const PATH_DIST = "dist/";
