--- lib/letter-avatar/template.ts
@@ -1,4 +1,0 @@
-
-export const TEMPLATE = `<ng-content></ng-content>`;
-
-
