--- lib/ngu-utility.ts
@@ -1,1 +1,0 @@
-export * from "./ngu-utility.module";
