+++ package.json
@@ -1,24 +1,26 @@
 {
   "name": "ngu-utility",
-  "version": "1.0.5",
+  "version": "1.0.6",
   "description": "Angular utility components,directives,pipes and services",
-  "main": "ngu-utility.umd.js",
-  "jsnext:main": "ngu-utility.esm.js",
-  "module": "ngu-utility.esm.js",
+  "main": "bundles/ngu-utility.umd.js",
+  "jsnext:main": "bundles/es5/ngu-utility.js",
+  "module": "bundles/es5/ngu-utility.js",
   "types": "ngu-utility.d.ts",
   "scripts": {
-    "copy": "node tools/copy-package.js && node tools/cleanup.js",
-    "bundle": "rimraf dist && rollup -c rollup.config.umd.js && rollup -c rollup.config.esm.js && tsc && ngc",
-    "postbundle": "node tools/removecode.js",
-    "build": "npm run bundle && npm run copy"
+    "build": "rimraf dist && ng-packagr -p package.json"
+  },
+  "ngPackage": {
+    "lib": {
+      "entryFile": "src/index.ts"
+    }
   },
   "keywords": [
     "angular",
     "angular2",
     "ngu-utility",
-	"letter-avatar",
-	"gmail avatar",
-	"letter avatar"
+    "letter-avatar",
+    "gmail avatar",
+    "letter avatar"
   ],
   "repository": {
     "type": "git",
@@ -47,12 +49,9 @@
     "@angular/platform-browser": "^5.0.0",
     "@angular/platform-browser-dynamic": "^5.0.0",
     "@types/node": "^7.0.14",
-    "del": "^2.2.2",
+    "ng-packagr": "^2.1.0",
     "node-sass": "^4.5.3",
-    "rollup-plugin-angular": "^0.4.4",
-    "rollup-plugin-commonjs": "^8.0.2",
-    "rollup-plugin-node-resolve": "^3.0.0",
-    "rollup-plugin-typescript": "^0.8.1",
+    "tsickle": "^0.27.2",
     "typescript": "^2.6.2"
   }
 }
