+++ src/index.ts
@@ -1,0 +1,1 @@
+export {NguUtilityModule} from "./ngu-utility.module";
