--- tools/copy-readme.js
@@ -1,4 +1,0 @@
-const fs = require('fs');
-
-let resizable = fs.readFileSync('README.md').toString();
-fs.writeFileSync('dist/README.md', resizable);
