+++ tsconfig.json
@@ -9,11 +9,5 @@
     "declaration": true,
     "outDir": "./dist",
     "lib": ["es2015", "dom"]
-  },
-  "files": [
-    "./lib/ngu-utility.ts"
-  ],
-  "angularCompilerOptions": {
-    "skipTemplateCodegen": true
   }
 }
